requirements.txt
torch==1.11.0
sklearn==0.0
scikit-learn==1.1.1
transformers==4.19.2
python-dotenv==0.20.0
numpy==1.22.4
matplotlib==3.5.2